Rearranged functions, added Groups interface
Added a Groups interface, along with some associated functions like destroy, getGroupsData, isPrivligeGroup, and some fields like cache. I also reorganized the code since some functions depend on other functions being declared before use. I was able fix more than 50% of the original errors, but I'm still stuck on a few of the remaining linter errors. One main issue I had is that I originally was working on a different file, but after going to office hours and working with a TA for quite some time, we decided it was best to move over to a different file because the original file I chose had one bug that no one could figure out and was the source of most of the 100+ errors. That said, even though I put a large amount of hours in the first file, I had to restart on this new file and ultimately struggled to fix all of the type errors.